horizon: rework shared font manager
